more cleanup

+++ src/main/java/com/hollingsworth/arsnouveau/client/gui/GuiManaHUD.java
@@ -48,7 +48,7 @@ public class GuiManaHUD {
         int offsetLeft = 10 + Config.MANABAR_X_OFFSET.get();
         int manaLength = 96;
 
-        manaLength *= (mana.getCurrentMana() / (double) (maxMana * (1.0 + ClientInfo.reservedOverlayMana)));
+        manaLength *= (mana.getCurrentMana() / (maxMana * (1.0 + ClientInfo.reservedOverlayMana)));
 
         int yOffset = minecraft.getWindow().getGuiScaledHeight() - 5 + Config.MANABAR_Y_OFFSET.get();
